moving demosaicing into its own files
