+++ src/fwd.h
@@ -8,19 +8,6 @@
 
 #include <nanogui/vector.h>
 
-
-
-
-inline nanogui::Vector3f operator*(const nanogui::Matrix3f & M, const nanogui::Vector3f & v)
-{
-	nanogui::Vector3f result;
-	result[0] = M.m[0][0] * v[0] + M.m[0][1] * v[1] + M.m[0][2] * v[2];
-	result[1] = M.m[1][0] * v[0] + M.m[1][1] * v[1] + M.m[1][2] * v[2];
-	result[2] = M.m[2][0] * v[0] + M.m[2][1] * v[1] + M.m[2][2] * v[2];
-	return result;
-}
-
-
 // forward declarations
 template <typename Vec, typename Value = typename Vec::Value, size_t Dims = Vec::Size> class Box;
 class Color3;
@@ -39,8 +26,7 @@ class ImageListPanel;
 class Timer;
 template<typename T> class Range;
 
-namespace nanogui
-{
+NAMESPACE_BEGIN(nanogui)
 class Widget;
 class Button;
 class CheckBox;
@@ -58,7 +44,7 @@ class MultiGraph;
 class Well;
 class ImageButton;
 class HSLGradient;
-}
+NAMESPACE_END(nanogui)
 
 // define some common types
 using Box2i     = Box<nanogui::Vector2i>;
